refactor: admin role can now see admin dashboard
